services/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
redis==5.2.1
asyncpg==0.30.0
PyJWT==2.9.0
python-multipart==0.0.9
